+++ src/hoc/layouts/Layout.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
-import SideDrawer from '../../Navigation/SideDrawer/SideDrawer';
-import NavBar from '../../Navigation/Navbar/Navbar';
+import SideDrawer from '../../components/Navigation/SideDrawer/SideDrawer';
+import NavBar from '../../components/Navigation/Navbar/Navbar';
 
 const MainWrapper = styled.div`
   width: 100%;
