+++ src/index.js
@@ -5,15 +5,19 @@ import theme from './utils/theme';
 import GlobalStyles from './utils/global';
 import { ThemeProvider } from 'styled-components';
 import { BrowserRouter } from 'react-router-dom';
+import { Provider } from 'react-redux';
+import store from './store';
 
 ReactDom.render(
-  <BrowserRouter>
-    <ThemeProvider theme={theme}>
-      <>
-        <App />
-        <GlobalStyles />
-      </>
-    </ThemeProvider>
-  </BrowserRouter>,
+  <Provider store={store}>
+    <BrowserRouter>
+      <ThemeProvider theme={theme}>
+        <>
+          <App />
+          <GlobalStyles />
+        </>
+      </ThemeProvider>
+    </BrowserRouter>
+  </Provider>,
   document.getElementById('root')
 );
