+++ src/store/reducers/authReducers.js
@@ -1,0 +1,5 @@
+const initialState = {};
+
+export default (state = initialState, action) => {
+  return 'something';
+};
