+++ src/store/reducers/index.js
@@ -1,0 +1,7 @@
+import { combineReducers } from 'redux';
+import authReducers from './authReducers';
+
+
+export default combineReducers({
+  auth: authReducers
+});
